Refactor shared left nav to render stats for a specified user
Before this change, the left nav always rendered info and stats for the
currently-logged-in user. Now, the view receives a user object to be
rendered. This way, the view can be re-used by the user profile view.

Additionally, if the user to be displayed is not the currently-logged-in
user, the content area and submit button for new twets are not rendered.